refactor!: remove deprecated GeneratedVaadinCheckbox
