+++ vaadin-checkbox-flow-parent/vaadin-checkbox-flow/src/main/java/com/vaadin/flow/component/checkbox/Checkbox.java
@@ -16,9 +16,19 @@
 package com.vaadin.flow.component.checkbox;
 
 import com.vaadin.flow.component.AbstractField;
+import com.vaadin.flow.component.AbstractSinglePropertyField;
+import com.vaadin.flow.component.ClickNotifier;
 import com.vaadin.flow.component.Component;
+import com.vaadin.flow.component.ComponentEvent;
+import com.vaadin.flow.component.ComponentEventListener;
+import com.vaadin.flow.component.Focusable;
 import com.vaadin.flow.component.HasLabel;
 import com.vaadin.flow.component.HasSize;
+import com.vaadin.flow.component.HasStyle;
+import com.vaadin.flow.component.Synchronize;
+import com.vaadin.flow.component.Tag;
+import com.vaadin.flow.component.dependency.JsModule;
+import com.vaadin.flow.component.dependency.NpmPackage;
 import com.vaadin.flow.component.html.Label;
 import com.vaadin.flow.component.shared.HasTooltip;
 import com.vaadin.flow.dom.PropertyChangeListener;
@@ -35,8 +45,14 @@ import com.vaadin.flow.dom.PropertyChangeListener;
  *
  * @author Vaadin Ltd
  */
-public class Checkbox extends GeneratedVaadinCheckbox<Checkbox, Boolean>
-        implements HasSize, HasLabel, HasTooltip {
+@Tag("vaadin-checkbox")
+@NpmPackage(value = "@vaadin/polymer-legacy-adapter", version = "24.0.0-alpha6")
+@JsModule("@vaadin/polymer-legacy-adapter/style-modules.js")
+@NpmPackage(value = "@vaadin/checkbox", version = "24.0.0-alpha6")
+@JsModule("@vaadin/checkbox/src/vaadin-checkbox.js")
+public class Checkbox extends AbstractSinglePropertyField<Checkbox, Boolean>
+        implements ClickNotifier<Checkbox>, Focusable<Checkbox>, HasLabel,
+        HasSize, HasStyle, HasTooltip {
 
     private final Label labelElement;
 
@@ -47,8 +63,7 @@ public class Checkbox extends GeneratedVaadinCheckbox<Checkbox, Boolean>
      * Default constructor.
      */
     public Checkbox() {
-        // initial value, default value, accept null value
-        super(false, false, false, true);
+        super("checked", false, false);
         getElement().addPropertyChangeListener("indeterminate",
                 "indeterminate-changed", NO_OP);
         getElement().addPropertyChangeListener("checked", "checked-changed",
@@ -211,9 +226,8 @@ public class Checkbox extends GeneratedVaadinCheckbox<Checkbox, Boolean>
      * @param autofocus
      *            the boolean value to set
      */
-    @Override
     public void setAutofocus(boolean autofocus) {
-        super.setAutofocus(autofocus);
+        getElement().setProperty("autofocus", autofocus);
     }
 
     /**
@@ -225,7 +239,7 @@ public class Checkbox extends GeneratedVaadinCheckbox<Checkbox, Boolean>
      * @return the {@code autofocus} property from the checkbox
      */
     public boolean isAutofocus() {
-        return isAutofocusBoolean();
+        return getElement().getProperty("autofocus", false);
     }
 
     /**
@@ -238,9 +252,8 @@ public class Checkbox extends GeneratedVaadinCheckbox<Checkbox, Boolean>
      *            the boolean value to set
      * @see #isIndeterminate()
      */
-    @Override
     public void setIndeterminate(boolean indeterminate) {
-        super.setIndeterminate(indeterminate);
+        getElement().setProperty("indeterminate", indeterminate);
     }
 
     /**
@@ -257,7 +270,27 @@ public class Checkbox extends GeneratedVaadinCheckbox<Checkbox, Boolean>
      *
      * @return the {@code indeterminate} property from the checkbox
      */
+    @Synchronize(property = "indeterminate", value = "indeterminate-changed")
     public boolean isIndeterminate() {
-        return isIndeterminateBoolean();
+        return getElement().getProperty("indeterminate", false);
+    }
+
+    /**
+     * If true, the user cannot interact with this element.
+     *
+     * @param disabled
+     *            the boolean value to set
+     */
+    protected void setDisabled(boolean disabled) {
+        getElement().setProperty("disabled", disabled);
+    }
+
+    /**
+     * If true, the user cannot interact with this element.
+     *
+     * @return the {@code disabled} property from the webcomponent
+     */
+    protected boolean isDisabledBoolean() {
+        return getElement().getProperty("disabled", false);
     }
 }
